Initial support for different inventory format
Created base class InventoryFormat
Moved implementation specific stuff to InventoryFormatOCS
