+++ Inventory.h
@@ -10,12 +10,8 @@
 
 #include <string>
 
-namespace tinyxml2 {
-    class XMLDocument;
-    class XMLElement;
-}
 
-class Machine;
+class InventoryFormat;
 class Inventory {
 public:
 	Inventory();
@@ -29,29 +25,8 @@ public:
 	bool Send(const char* serverUrl);
 	void Print();
 
-	int Checksum() const;
-
 private:
-    void _AddAccountInfo();
-    void _AddBIOSInfo();
-    void _AddCPUsInfo();
-    void _AddStoragesInfo();
-    void _AddMemoriesInfo();
-    void _AddDrivesInfo();
-    void _AddHardwareInfo();
-    void _AddNetworksInfo();
-    void _AddProcessesInfo();
-    void _AddSoftwaresInfo();
-    void _AddUsersInfo();
-    void _AddVideosInfo();
-    void _AddMonitorsInfo();
-
-    bool _WriteProlog(tinyxml2::XMLDocument& document) const;
-
-    std::string _GenerateDeviceID() const;
-
-    tinyxml2::XMLDocument* fDocument;
-    tinyxml2::XMLElement* fContent;
+	InventoryFormat* fFormat;
 };
 
 #endif /* INVENTORY_H_ */
