+++ inventoryformat/InventoryFormat.cpp
@@ -1,0 +1,18 @@
+/*
+ * InventoryFormat.cpp
+ *
+ *  Created on: 31 ago 2022
+ *      Author: Stefano Ceccherini
+ */
+
+#include "../inventoryformat/InventoryFormat.h"
+
+InventoryFormat::InventoryFormat()
+{
+}
+
+
+/* virtual */
+InventoryFormat::~InventoryFormat()
+{
+}
